refactor(BE): extract files related to global auth from /api/auth to /auth
