+++ backend/src/auth/strategies/jwt.strategy.ts
@@ -2,7 +2,7 @@ import { Injectable } from '@nestjs/common'
 import { PassportStrategy } from '@nestjs/passport'
 import { ExtractJwt, Strategy } from 'passport-jwt'
 
-import type { JWTPayload } from './auth.interface'
+import type { JWTPayload } from '../auth.interface'
 
 @Injectable()
 export class JwtStrategy extends PassportStrategy(Strategy) {
